+++ lib/rackla/rackla.ex
@@ -1,5 +1,5 @@
 defmodule Rackla do
-  @moduledoc Regex.replace(~r/```(elixir|json)(\n|.*)```/rs, File.read!("README.md"), 
+  @moduledoc Regex.replace(~r/```(elixir|json)(\n|.*)```/Us, File.read!("README.md"), 
   fn(_, _, code) -> Regex.replace(~r/^/m, code, "    ") end)
   
   import Plug.Conn
@@ -33,6 +33,9 @@ defmodule Rackla do
    default: `false`.
    * `:follow_redirect` - If set to true, Rackla will follow redirects, 
    default: `false`.
+   * `:max_redirect` - Maximum number of redirects, default: `5`.
+   * `:force_redirect` - Force follow redirect (e.g. POST), default: `false`.
+   * `:proxy` - Proxy to use, see `Rackla.Proxy`, default: `nil`.
    
   If you specify any options in a `Rackla.Request` struct, these will overwrite
   the options passed to the `request` function for that specific request.
@@ -43,7 +46,12 @@ defmodule Rackla do
   def request(requests, options) when is_list(requests) do
     producers =
       Enum.map(requests, fn(request) ->
-        if is_binary(request), do: request = %Rackla.Request{url: request}
+        request = 
+          if is_binary(request) do 
+            %Rackla.Request{url: request}
+          else
+            request
+          end
 
         {:ok, producer} =
           Task.start_link(fn ->
@@ -52,6 +60,45 @@ defmodule Rackla do
             global_connect_timeout = Keyword.get(options, :connect_timeout, 5_000)
             global_receive_timeout = Keyword.get(options, :receive_timeout, 5_000)
             global_follow_redirect = Keyword.get(options, :follow_redirect, false)
+            global_max_redirect = Keyword.get(options, :max_redirect, 5)
+            global_force_redirect = Keyword.get(options, :force_redirect, false)
+            
+            global_proxy = Keyword.get(options, :proxy)
+            request_proxy = Map.get(request_options, :proxy)
+            
+            rackla_proxy = 
+              cond do
+                request_proxy -> request_proxy
+                global_proxy -> global_proxy
+                true -> nil
+              end
+            
+            proxy_options =
+              case rackla_proxy do
+                %Rackla.Proxy{type: type, host: host, port: port, username: username, password: password, pool: pool} ->
+                  proxy_basic_setting = [proxy: {type, String.to_char_list(host), port}]
+                  
+                  auth_settings =
+                    case type do
+                      :socks5 ->
+                        socks5_user = if username, do: [socks5_user: username], else: []
+                        socks5_pass = if password, do: [socks5_pass: password], else: []  
+                        
+                        socks5_user ++ socks5_pass
+                      :connect -> 
+                        if username && password do
+                          [proxy_auth: {username, password}]
+                        else
+                         []
+                       end
+                    end
+                  
+                  pool_setting = if pool, do: [pool: pool], else: []
+                  
+                  proxy_basic_setting ++ auth_settings ++ pool_setting
+                
+                nil -> []
+              end
 
             hackney_request =
               :hackney.request(
@@ -63,11 +110,16 @@ defmodule Rackla do
                   insecure: Map.get(request_options, :insecure, global_insecure),
                   connect_timeout: Map.get(request_options, :connect_timeout, global_connect_timeout),
                   recv_timeout: Map.get(request_options, :receive_timeout, global_receive_timeout),
-                  follow_redirect: Map.get(request_options, :follow_redirect, global_follow_redirect)
-                ]
+                  follow_redirect: Map.get(request_options, :follow_redirect, global_follow_redirect),
+                  max_redirect: Map.get(request_options, :max_redirect, global_max_redirect),
+                  force_redirect: Map.get(request_options, :force_redirect, global_force_redirect)
+                ] ++ proxy_options
               )
 
             case hackney_request do
+              {:ok, {:maybe_redirect, _, _, _}} -> 
+                warn_request(:force_redirect_disabled)
+              
               {:ok, status, headers, body_ref} ->
                 case :hackney.body(body_ref) do
                   {:ok, body} ->
@@ -410,6 +462,83 @@ defmodule Rackla do
         response_async(rackla, conn, options)
     end
   end
+  
+  @doc """
+  Convert an incoming request (from `Plug`) to a `Rackla.Request`.
+  If `options` is specified, it will be added to the `Rackla.Request`.
+  For valid options, see documentation for `Rackla.Request`.
+  
+  Returns either `{:ok, Rackla.Request}` or `{:error, reason}` as per 
+  `:gen_tcp.recv/2`.
+  
+  The `Plug.Conn` will be taken implicitly by looking for a variable named 
+  `conn`. If you want to specify which `Plug.Conn` to use, you can use 
+  `Rackla.incoming_request_conn`.
+  
+  Using this macro is the same as writing:
+    `conn = incoming_request_conn(conn, options)`
+  
+  From `Plug.Conn` documentation:
+  Because the request body can be of any size, reading the body will only work 
+  once, as Plug will not cache the result of these operations. If you need to 
+  access the body multiple times, it is your responsibility to store it. Finally 
+  keep in mind some plugs like Plug.Parsers may read the body, so the body may 
+  be unavailable after being accessed by such plugs.
+  """
+  @spec incoming_request(%{}) :: {:ok, Rackla.Request.t} | {:error, atom}
+  defmacro incoming_request(options \\ %{}) do
+    quote do
+      {var!(conn), rackla_request} = incoming_request_conn(var!(conn), unquote(options))
+      _ = var!(conn) # hack to get rid of "unused variable" compiler warning
+      rackla_request
+    end
+  end
+    
+  @doc """
+  See documentation for `Rackla.incoming_request`.
+  """
+  @spec incoming_request_conn(Plug.Conn.t, %{}) :: {Plug.Conn.t, {:ok, Rackla.Request.t}} | {Plug.Conn.t, {:error, atom}}
+  def incoming_request_conn(conn, options \\ %{}) do
+    response_body = 
+      Stream.unfold(Plug.Conn.read_body(conn), fn 
+        :done -> 
+          nil;
+        
+        {:ok, body, new_conn} -> 
+          {{new_conn, body}, :done};
+        
+        {:more, partial_body, new_conn} -> 
+          {partial_body, Plug.Conn.read_body(new_conn)};
+        
+        {:error, term} -> 
+          {{:error, term}, :done}
+      end)
+      |> Enum.reduce({"", conn}, fn 
+        ({:error, term}, {_body_acc, conn_acc}) -> {{:error, term}, conn_acc};
+        ({new_conn, body}, {body_acc, _conn_acc}) -> {{:ok, body_acc <> body}, new_conn};
+        (partial_body, {body_acc, conn_acc}) -> {body_acc <> partial_body, conn_acc}
+      end)
+      
+    case response_body do
+      {{:error, term}, final_conn} -> {final_conn, {:error, term}}
+      
+      {{:ok, body}, final_conn} ->
+        method = conn.method |> String.downcase |> String.to_atom
+        url = "#{Atom.to_string(conn.scheme)}://#{conn.host}#{conn.request_path}"
+        headers = Enum.into(conn.req_headers, %{})
+        
+        rackla_request =
+          %Rackla.Request{
+            method: method,
+            url: url,
+            headers: headers,
+            body: body,
+            options: options
+          }
+          
+        {final_conn, {:ok, rackla_request}}
+    end
+  end
 
   @spec response_async(t, Plug.Conn.t, Keyword.t) :: Plug.Conn.t
   defp response_async(%Rackla{} = rackla, conn, options) do
@@ -433,7 +562,7 @@ defmodule Rackla do
   defp send_chunks(producers, conn) when is_list(producers) do
     send_thing = 
       fn(thing, remaining_producers, conn) ->
-        unless is_binary(thing), do: thing = inspect(thing)
+        thing = if is_binary(thing), do: thing, else: inspect(thing)
 
         case chunk(conn, thing) do
           {:ok, new_conn} ->
@@ -488,8 +617,8 @@ defmodule Rackla do
           response_sync_chunk(nested_rackla, conn, options)
           
         {^pid, thing} ->
-          if elem(thing, 0) == :ok, do: thing = elem(thing, 1)
-          unless is_binary(thing), do: thing = inspect(thing)
+          thing = if elem(thing, 0) == :ok, do: elem(thing, 1), else: thing
+          thing = if is_binary(thing), do: thing, else: inspect(thing)
           
           case chunk(conn, thing) do
             {:ok, new_conn} ->
@@ -544,24 +673,31 @@ defmodule Rackla do
         headers = Keyword.get(options, :headers, %{})
         compress = Keyword.get(options, :compress, false)
         
-        if compress do
-          allow_gzip = 
-            Plug.Conn.get_req_header(conn, "accept-encoding")
-            |> Enum.flat_map(fn(encoding) ->
-              String.split(encoding, ",", trim: true)
-              |> Enum.map(&String.strip/1)
-            end)
-            |> Enum.any?(&(Regex.match?(~r/(^(\*|gzip)(;q=(1$|1\.0{1,3}$|0\.[1-9]{1,3}$)|$))/, &1)))
-          
-          if allow_gzip || compress == :force do
-            response_binary = :zlib.gzip(response_binary)
-            headers = Map.merge(headers, %{"content-encoding" => "gzip"})
-          end
-        end
+        {response_binary, headers} = 
+          if compress do
+            allow_gzip = 
+              Plug.Conn.get_req_header(conn, "accept-encoding")
+              |> Enum.flat_map(fn(encoding) ->
+                String.split(encoding, ",", trim: true)
+                |> Enum.map(&String.strip/1)
+              end)
+              |> Enum.any?(&(Regex.match?(~r/(^(\*|gzip)(;q=(1$|1\.0{1,3}$|0\.[1-9]{1,3}$)|$))/, &1)))
 
-        if Keyword.get(options, :json, false) do
-          conn = put_resp_content_type(conn, "application/json")
-        end
+            if allow_gzip || compress == :force do
+              {:zlib.gzip(response_binary), Map.merge(headers, %{"content-encoding" => "gzip"})}
+            else
+              {response_binary, headers}
+            end
+          else
+            {response_binary, headers}
+          end
+
+        conn = 
+          if Keyword.get(options, :json, false) do
+            put_resp_content_type(conn, "application/json")
+          else
+            conn
+          end
 
         chunk_status =
           prepare_conn(conn, Keyword.get(options, :status, 200), headers)