+++ lib/rackla/router_gen_server.ex
@@ -9,7 +9,7 @@ defmodule Rackla.RouterGenServer do
 
   def init(_args) do
     port = Application.get_env(:rackla, :port, 4000)
-    if is_binary(port), do: port = String.to_integer(port)
+    port = if is_binary(port), do: String.to_integer(port), else: port
 
     Plug.Adapters.Cowboy.http(Router, [], port: port)
   end